Code as of 2015-08-19

+++ include/Control.h
@@ -6,7 +6,7 @@ void updateLoopOnce();
 
 void setupMotors();
 
-void setMotors(uint8_t ldem, uint8_t rdem);
+void setMotors(int16_t ldem, int16_t rdem);
 
 void calculateLocation();
